Fixed Adding a basket

+++ BasketService/Models/Basket.cs
@@ -24,7 +24,7 @@ namespace BasketService
         public long Id { get; set; }
         public string? Title { get; set; }
         public string? Description { get; set; }
-        public decimal? Price { get; set; }
+        public float? Price { get; set; }
         public long? Quantity { get; set; }
 
         [JsonIgnore]
